refactor: keep delivery modules within review limits
