+++ src/model_routing.rs
@@ -975,24 +975,20 @@ pub(crate) use state_routing::{
 };
 
 #[cfg(test)]
-#[path = "model_routing_tests.rs"]
-pub(crate) mod tests;
-
-#[cfg(test)]
 #[path = "model_routing_health_tests.rs"]
 mod health_tests;
-
-#[cfg(test)]
-#[path = "model_routing_snapshot_tests.rs"]
-mod snapshot_tests;
-
 #[cfg(test)]
 #[path = "model_routing_pool_tests.rs"]
 mod pool_tests;
-
 #[cfg(test)]
 #[path = "model_routing_recovery_tests.rs"]
 mod recovery_tests;
+#[cfg(test)]
+#[path = "model_routing_snapshot_tests.rs"]
+mod snapshot_tests;
+#[cfg(test)]
+#[path = "model_routing_tests.rs"]
+pub(crate) mod tests;
 
 #[cfg(test)]
 #[path = "model_routing_evidence_tests.rs"]
